requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
mistralai==0.0.10
pinecone-client==2.2.4
langchain==0.1.0
langchain-community==0.0.13
langchain-mistralai==0.0.2
pydantic==1.3.1
python-dotenv==1.0.0
